backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
httpx==0.27.2
pillow==10.4.0
python-multipart==0.0.9
slowapi==0.1.9
pydantic==2.9.0
numpy==1.26.4
stripe==11.3.0
google-generativeai==0.8.6
python-dotenv==1.0.0
PyJWT==2.8.0
easyocr==1.7.0
ffmpeg-python==0.2.0
torch>=2.0.0
opencv-python>=4.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
